+++ src/widgets/editor/TopControls.tsx
